+++ packages/flow_widget_macos/macos/flow_widget_macos/Package.swift
@@ -1,0 +1,21 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+  name: "flow_widget_macos",
+  platforms: [
+    .macOS("11.0")
+  ],
+  products: [
+    .library(name: "flow-widget-macos", targets: ["flow_widget_macos"])
+  ],
+  dependencies: [],
+  targets: [
+    .target(
+      name: "flow_widget_macos",
+      dependencies: []
+    )
+  ]
+)